backend/requirements.txt
# backend/requirements.txt

flask
torch
pillow
ultralytics
opencv-python-headless
flask_cors